pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "resumakeai"
version = "0.8.0"
description = "Build styled CV documents from a single YAML source."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Fayssal El Mofatiche" }]
keywords = ["cv", "resume", "docx", "yaml", "cli", "word", "document"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup",
    "Typing :: Typed",
]
dependencies = [
    "python-docx>=1.1.0",
    "pyyaml>=6.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
ai = ["resumakeai[anthropic]"]
pdf = ["docx2pdf>=0.1.8"]
weasyprint = ["weasyprint>=62.0"]
linkedin = ["pdfplumber>=0.10.0"]
watch = ["watchdog>=4.0.0"]
all = ["resumakeai[anthropic,openai,pdf,weasyprint,linkedin,watch]"]
dev = ["pytest>=8.0", "ruff>=0.8.0"]

[project.urls]
Homepage = "https://github.com/fayssal-elmofatiche/resumake"
Repository = "https://github.com/fayssal-elmofatiche/resumake"
Issues = "https://github.com/fayssal-elmofatiche/resumake/issues"
Changelog = "https://github.com/fayssal-elmofatiche/resumake/blob/main/CHANGELOG.md"

[project.scripts]
resumake = "resumake:app"

[tool.setuptools.packages.find]
include = ["resumake"]

[tool.setuptools.package-data]
resumake = ["assets/*.png", "themes/*.yaml", "templates/*.yaml", "py.typed"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"resumake/__init__.py" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
